% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{bdmep_coords}
\alias{bdmep_coords}
\title{Get coordinates on meteorological stations}
\usage{
bdmep_coords()
}
\value{
A data frame is returned with metadata, including the stations
 \code{id}, and coordinates (\code{lon}, \code{lat}, \code{alt})
}
\description{
Fetch the coordinates on meteorological stations from INMET
\url{http://www.inmet.gov.br/projetos/rede/pesquisa/lista_estacao.php}
}
\examples{

info <- bdmep_coords()
head(info)
}
\author{
Jonatan Tatsch
}
