% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bdmep_data_status}
\alias{bdmep_data_status}
\title{Report status of each variable}
\usage{
bdmep_data_status(data_bdmep = xtidy)
}
\arguments{
\item{data_bdmep}{data processed by \code{\link{bdmep_read}} in 
\code{\link{bdmep_import_station}}.}
}
\value{
data frame with the percentage of valid observations for each variable
 \describe{
   \item{id}{station id}
   \item{sdate}{start date of observations}
   \item{edate}{end date of observations}
   \item{rows}{number of rows in data file}
   \item{request_status}{character scalar with information on the status of a request}
   \item{prec}{valid observations of prec in percentage}
   \item{...}{valid observations of ith variable in percentage}
   \item{ws}{valid observations of ws in percentage}
 }
}
\description{
Report status of each variable
}
