% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_import}
\alias{bdmep_import}
\title{Import data from Brazilian meteorological stations}
\usage{
bdmep_import(ids = c("83936", "83967"), sdate = "01/01/1961",
  edate = format(Sys.Date(), "\%d/\%m/\%Y"), email = "your-email",
  passwd = "your-password", verbose = TRUE)
}
\arguments{
\item{ids}{a character vector with codes of meteorological stations}

\item{sdate}{start date in "d/m/Y" format}

\item{edate}{end date in "d/m/Y" format, default values \code{format(Sys.Date(), "\%d/\%m/\%Y")}}

\item{email}{e-mail to access BDMEP}

\item{passwd}{password to access BDMEP}

\item{verbose}{if TRUE, prints login sucessfull; if not, not. Default is TRUE.}
}
\value{
a data frame with variables in columns ((see \code{\link{bdmep_description}})) and observations (date and time) along rows.
}
\description{
Import data from Brazilian meteorological stations
}
\details{
The data are in sub-daily time scale. A minimum data quality control is applied to the data.
This include: a chronological sequence check; filling data from missing dates with NA; 
remove duplicated data. Time variables (year, month, day, hour) are aggregated into a POSIX object in UTC
}
\examples{
# download data for Santa Maria and Porto Alegre
metdata <- bdmep_import(ids = c("83936", "83967"), 
                        sdate = "01/01/1961",
                        edate = format(Sys.Date(), '\%d/\%m/\%Y'),
                        email = "your-email",
                        passwd = "your-password",
                        verbose = TRUE)
head(metdata)
tail(metdata)
summary(metdata)

}
\author{
Jonatan Tatsch
}
