% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_info}
\alias{bdmep_info}
\title{Get OMM code, state and station name on meteorological stations from INMET}
\usage{
bdmep_info()
}
\arguments{
\item{none}{\url{http://www.inmet.gov.br/projetos/rede/pesquisa/lista_estacao.php}.}
}
\value{
A data frame is returned with 
 \code{nome}, \code{estado}, \code{codigo_omm}
}
\description{
Get basic information on meteorological station from INMET
}
\details{
This function is used to find the OMM code that can be
used to import BDMEP data using \code{\link{import_bdmep}}.
}
\examples{

\dontrun{
# this can take a while
metad <- get_meta_bdmep()
head(metad, 15)
#save(metad, file = "data/metad.rda")
# plot locations
with(metad, plot(lon, lat, pch = 4))
}
}
\author{
Jônatan Tatsch
}

