% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_login_att}
\alias{bdmep_login_att}
\title{Get login attributes form to acces BDMEP}
\usage{
bdmep_login_att(lnk, email, passwd)
}
\arguments{
\item{lnk}{url to BDMEP}
}
\value{
a named list with user name, password and text of button access
}
\description{
Get login attributes form to acces BDMEP
}
\author{
Jonatan Tatsch
}
