% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{bdmep_meta}
\alias{bdmep_meta}
\title{Metadata on 394 meteorological stations of INMET}
\format{A data frame with 394 rows and 7 columns.
	\describe{
		\item{id}{station id, character with 5 digits corresponding to OMM code}
		\item{lon}{longitude in decimal degrees}
		\item{lat}{latitude in decimal degrees}
		\item{alt}{altitude in metres}
		\item{name}{station name}
		\item{state}{Full name of state}
		\item{uf}{Federative Unit}
}}
\source{
\code{bdmep_meta} was compiled from 
\url{http://www.inmet.gov.br/webcdp/climatologia/normais/imagens/normais/planilhas/Relac_Est_Meteo_NC.xls}
}
\description{
Spatial coordinates, altitude and others informations of INMET's 
meteorological stations.
}
\details{
\code{bdmep_meta} is supplied with the \code{inmetr} package and
consists of the following information.
}
\keyword{datasets}
