% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{bdmep_metadata_normclim}
\alias{bdmep_metadata_normclim}
\title{Get metadata on meteorological stations from alternative link to climate stations from INMET}
\usage{
bdmep_metadata_normclim(metadata_file = "data/relac_est_meteo_nc.RData",
  verbose = TRUE)
}
\value{
data frame with variables:
 \code{id}, \code{lon}, \code{lat}, \code{alt}, 
 \code{name}, \code{state}, \code{uf}.
 @details First the function try read metadata in ./data/relac_est_meteo_nc.rds
}
\description{
Get metadata on meteorological stations from alternative link to climate stations from INMET
}
\examples{
 metadt <- bdmep_metadata_normclim()
 head(metadt)
}
