% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_read}
\alias{bdmep_read}
\title{Read data downaloaded from BDMEP}
\usage{
bdmep_read(x)
}
\arguments{
\item{x}{a numeric vector with the meteorological station code}
}
\value{
a data frame with variables in columns and observations along rows
}
\description{
Read and tidy data downloaded with \code{\link{bdmep_import}}
}
\details{
A minimum quality control check is applied to the data
This include: a chronological sequence check; filling missing dates with NA; 
remove duplicated data; aggregate time and date information into a POSIX object
}
\author{
Jonatan Tatsch
}
