% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_stations}
\alias{bdmep_stations}
\title{Get OMM code, state and station name on meteorological stations from INMET
\url{http://www.inmet.gov.br/projetos/rede/pesquisa/lista_estacao.php}.}
\usage{
bdmep_stations()
}
\value{
a data frame is returned with 
 \code{name}, \code{state}, \code{id}
}
\description{
Get basic information on meteorological station from INMET
}
\details{
This function is used to find the OMM code that can be
used to import BDMEP data using \code{\link{import_bdmep}}.
}
\examples{

\dontrun{
# tyr get information from inmet site
stns_id <- bdmep_stations()
head(stns_id, 15)
#save(stns_id, file = "data/stns_id.rda")
}
}
\author{
Jonatan Tatsch
}

