% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_template}
\alias{bdmep_template}
\title{Template bdmep dataframe to be used when the status of a request was not successfully executed.}
\usage{
bdmep_template(id, req_msg)
}
\arguments{
\item{id}{a character scalar with the meteorological station code}

\item{req_msg}{character scalar with information on the status of a request}
}
\value{
a dataframe with values filled with NA, except for id and request_status
}
\description{
Template bdmep dataframe to be used when the status of a request was not successfully executed.
}
\examples{
bdmep_template(id = "83936", req_msg = "Bad Gateway (HTTP 502).")
}
