% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_units}
\alias{bdmep_units}
\title{Units of meteorological variables}
\usage{
bdmep_units()
}
\value{
a data frame is returned with 
 \code{varname}, \code{description}, \code{unit}
}
\description{
Get variable names, description and units
}
\details{
This function describe the Meteorological variables imported with \code{\link{bdmep_import}}

to information about instruments see \url{http://www.inmet.gov.br/portal/index.php?r=home/page&page=instrumentos}
}
\examples{
met_vars <- bdmep_units()
met_vars

}
\author{
Jonatan Tatsch
}
