% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{clean_col_names}
\alias{clean_col_names}
\title{Clean column names of a data frame obtained a excel file}
\usage{
clean_col_names(x)
}
\arguments{
\item{x}{data frame, tibble or data_frame}
}
\value{
data frame with names id, nome, estado, uf, lat, lon, alt
}
\description{
Clean column names of a data frame obtained a excel file
}
\details{
This functions is used in \code{\link{read_meteo_stn_xls}}
}
