% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{data_description}
\alias{data_description}
\title{Meteorological variables description}
\usage{
data_description()
}
\value{
a data frame is returned with 
 \code{varname}, \code{description}, \code{unit}
}
\description{
Get variable names, description and units
}
\details{
This function describe the Meteorological variables imported with \code{\link{import_bdmep}}

to information about instruments see \url{http://www.inmet.gov.br/portal/index.php?r=home/page&page=instrumentos}
}
\examples{
met_vars <- data_description()
met_vars

}
\author{
Jonatan Tatsch
}
