% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{get_meta_bdmep}
\alias{get_meta_bdmep}
\title{Get OMM code and other meta data on meteorological stations from INMET}
\usage{
get_meta_bdmep()
}
\arguments{
\item{none}{\url{http://www.inmet.gov.br/projetos/rede/pesquisa/lista_estacao.php}.}
}
\value{
A data frame is returned with meta data, including a
 \code{site} that can be supplied to
\code{\link{import_bdmep}}, coordinates \code{lon}, \code{lat}, and
 \code{alt}.
}
\description{
Get information on conventional meteorological station from INMET
}
\details{
This function is used to find the OMM code that can be
used to import BDMEP data using \code{\link{import_bdmep}}.
}
\examples{

\dontrun{
# this can take a while
metad <- get_meta_bdmep()
head(metad, 15)
#save(metad, file = "data/metad.rda")
# plot locations
with(metad, plot(lon, lat, pch = 4))
}
}
\author{
Jônatan Tatsch
}

