% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{import_bdmep}
\alias{import_bdmep}
\title{Import data from BDMEP-INMET site}
\usage{
import_bdmep(id = "83586", sdate = "01/01/1961",
  edate = format(Sys.Date(), "\%d/\%m/\%Y"), email = "your-email",
  passwd = "your-password")
}
\arguments{
\item{id}{a numeric vector with the meteorological station code}

\item{sdate}{start date in "d/m/Y" format}

\item{edate}{end date in "d/m/Y" format, default valueis \code{format(Sys.Date(), "\%d/\%m/\%Y")}}

\item{email}{e-mail to access BDMEP}

\item{passwd}{password to access BDMEP}
}
\value{
a data frame with variables in columns and observations along rows
}
\description{
Import data from \url{http://www.inmet.gov.br/projetos/rede/pesquisa}
}
\details{
A minimum quality control check is applied to the data.
This include: a chronological sequence check; filling missing dates with NA; 
remove duplicated data. Time variables (year, month, day, hour) are aggregated into a POSIX object in UTC
}
\examples{
# download data for Santa Maria-RS 
sm <- import_bdmep(id = 83936, email = "your-email", passwd = "your-password")
head(sm)
summary(sm)

}
\author{
Jonatan Tatsch
}

