% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inmetr.R
\docType{package}
\name{inmetr}
\alias{inmetr}
\title{inmetr:A package for import data from Brazilian meteorological stations}
\description{
The inmetr package offers a set of functions to import historical data from Brazilian meteorological stations available in the Meteorological Database for Education and Research \href{http://www.inmet.gov.br/projetos/rede/pesquisa}{BDMEP} of National Institute of Meteorology \href{http://www.inmet.gov.br}{INMET}
}
\section{inmetr functions}{

The main functions are:
\itemize{
   \item \code{\link{bdmep_import}}
   \item \code{\link{bdmep_description}}
}
 The data included in \code{inmetr} is:
\itemize{
   \item \code{\link{bdmep_meta}}
}
}

