% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{parse_coords}
\alias{parse_coords}
\title{Convert a character string with information of spatial coordinates (lon, lat) to numeric}
\usage{
parse_coords(x)
}
\arguments{
\item{x}{character vector}
}
\value{
numeric vector
}
\description{
Convert a character string with information of spatial coordinates (lon, lat) to numeric
}
\details{
This function is used to clean data in \code{\link{read_meteo_stn_xls}}
}
