% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{read_bdmep}
\alias{read_bdmep}
\title{Read BDMEP data file}
\usage{
read_bdmep(x)
}
\arguments{
\item{x}{a numeric vector with the meteorological station code}
}
\value{
a data frame with variables in columns and observations along rows
}
\description{
Read and tidy data downloaded with \code{\link{import_bdmep}}
}
\details{
A minimum quality control check is applied to the data
This include: a chronological sequence check; filling missing dates with NA; 
remove duplicated data; aggregate time information into a POSIX object
}
\author{
Jonatan Tatsch
}

