% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_meteo_stn_xls}
\alias{read_meteo_stn_xls}
\title{Read excel file with metadata of INMET's climate stations.}
\usage{
read_meteo_stn_xls(file = "./inst/extdata/relac_est_meteo_nc.xls",
  verbose = FALSE)
}
\arguments{
\item{file}{path to excel file download from alternative link to climate stations from INMET}

\item{verbose}{logical, if TRUE show messages.}
}
\value{
data frame with variables \code{id}, \code{lon}, \code{lat}, \code{alt}, \code{nome}, \code{estado}, \code{uf}.
}
\description{
Read excel file with metadata of INMET's climate stations.
}
\details{
The \code{file} is download from \url{http://www.inmet.gov.br/webcdp/climatologia/normais/imagens/normais/planilhas/Relac_Est_Meteo_NC.xls}.
This function is used in \code{\link{bdmep_metadata_normclim}}.
}
