% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{set_bdmep_user}
\alias{set_bdmep_user}
\title{Set username and password for the BDMEP access}
\usage{
set_bdmep_user(lnk, email, passwd)
}
\arguments{
\item{lnk}{url to BDMEP access}

\item{email}{your BDMEP username}

\item{passwd}{your BDMEP password}
}
\value{
a named list with user name, password and text of button access
}
\description{
Set username and password for the BDMEP access
}
\author{
Jonatan Tatsch
}
