% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_empty}
\alias{str_empty}
\title{Detect if a string is empty}
\usage{
str_empty(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}
}
\value{
logical, TRUE in the absence of a string (""), otherwise FALSE.
}
\description{
Detect if a string is empty
}
