% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_import}
\alias{bdmep_import}
\title{Import data from Brazilian meteorological stations}
\usage{
bdmep_import(id = c("83936", "83967"), sdate = "01/01/1961",
  edate = format(Sys.Date(), "\%d/\%m/\%Y"), email = "your-email",
  passwd = "your-password", verbose = TRUE, destdir = NULL, ...)
}
\arguments{
\item{id}{A character vector with codes of meteorological stations}

\item{sdate}{Start date in "d/m/Y" format}

\item{edate}{End date in "d/m/Y" format, default values \code{format(Sys.Date(), "\%d/\%m/\%Y")}}

\item{email}{E-mail to access BDMEP}

\item{passwd}{Password to access BDMEP}

\item{verbose}{If TRUE, prints login sucessfull.}

\item{destdir}{A character string with the path where the downloaded data is saved. If it is  NULL, data will not be saved in disk.}

\item{...}{Additional arguments for the underlying export functions (see \code{\link{export}}).}
}
\value{
A data frame with variables in columns (see \code{\link{bdmep_description}}) and observations (date and time) along rows.
}
\description{
Import data from Brazilian meteorological stations
}
\details{
The data are in sub-daily time scale. A minimum data quality control is applied to the data.
This include: a chronological sequence check; filling data from missing dates with NA; 
remove duplicated data. Time variables (year, month, day, hour) are aggregated into a POSIX object in UTC
}
\examples{
# download data for Santa Maria and Porto Alegre
metdata <- bdmep_import(id = c("83936", "83967"), 
                        sdate = "01/01/1961",
                        edate = format(Sys.Date(), '\%d/\%m/\%Y'),
                        email = "your-email",
                        passwd = "your-password",
                        verbose = TRUE)
head(metdata)
tail(metdata)
summary(metdata)

}
\author{
Jonatan Tatsch
}
