% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{bdmep_metadata}
\alias{bdmep_metadata}
\title{Get metadata on meteorological stations}
\usage{
bdmep_metadata()
}
\value{
A data frame is returned with metadata, including the stations
 \code{id}, and coordinates (\code{lon}, \code{lat}, \code{alt})
}
\description{
Fetch the coordinates on meteorological stations from INMET
\url{http://www.inmet.gov.br/projetos/rede/pesquisa/lista_estacao.php}
}
\examples{

info <- bdmep_metadata()
head(info)
}
\author{
Jonatan Tatsch
}
