% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inmetr.R
\docType{package}
\name{inmetr}
\alias{inmetr}
\alias{inmetr-package}
\title{inmetr:	A	package	for	import data from Brazilian meteorological stations}
\description{
The inmetr package offers a set of functions to import historical data from Brazilian meteorological stations available in the Meteorological Database for Education and Research \href{http://www.inmet.gov.br/projetos/rede/pesquisa}{BDMEP} of National Institute of Meteorology \href{http://www.inmet.gov.br}{INMET}
}
\section{inmetr functions}{

The main functions are:
\itemize{
   \item \code{\link{import_bdmep}}  
   \item \code{\link{bdmep_stations}} 
   \item \code{\link{bdmep_metadata}} 
   \item \code{\link{data_description}} 
}
}

